refactor: changed mysql usage to sequelize and change the code structure
