+++ package.json
@@ -27,6 +27,7 @@
 		"helmet": "^7.1.0",
 		"jsonwebtoken": "^9.0.2",
 		"mysql2": "^3.9.7",
+		"sequelize": "^6.37.3",
 		"uuid": "^9.0.1"
 	},
 	"devDependencies": {
